requirements.txt
google-genai
qdrant-client
pdfplumber
pandas
openpyxl
msal
requests
python-dotenv
